(1.3.1) Add NAMESPACE.
Fix R CMD check failure on test. The nature of this was how the variance of an entire vector of NA values is computed (R throws an error where in the past it returned NA), the behavior of rowVar() and colVar() on BufferedMatrix objects remains unchanged (that is they return NA).


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/BufferedMatrix@30030 bc3139a8-67e5-0310-9ffc-ced21a209358